[Feat] 전략패턴에 맞게 Notification 로직 수정
