+++ src/main/java/com/tools/seoultech/timoproject/notification/NotificationController.java
@@ -2,6 +2,9 @@ package com.tools.seoultech.timoproject.notification;
 
 import java.util.List;
 
+import com.tools.seoultech.timoproject.notification.dto.NotificationRequest;
+import com.tools.seoultech.timoproject.notification.dto.NotificationResponse;
+import com.tools.seoultech.timoproject.notification.service.NotificationService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
